Implemented Transaction/TransactionDto and other related services classes

+++ src/main/java/hu/csercsak_albert/banking_system/repository/TransactionRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface TransactionRepository extends JpaRepository<Transaction, Long> {
 
-    List<Transaction> findByFromAccountId(int fromAccountId);
+    List<Transaction> findByFromAccountId(long fromAccountId);
 
-    List<Transaction> findByToAccountId(int toAccountId);
+    List<Transaction> findByToAccountId(long toAccountId);
 }
